pull master

+++ src/main/java/com/doudizu/seckill/dao/OrderDao.java
@@ -4,13 +4,14 @@ import com.doudizu.seckill.domain.Order;
 import com.doudizu.seckill.domain.OrderInfo;
 import org.apache.ibatis.annotations.*;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Mapper
 public interface OrderDao {
     //插入订单
     @Insert("insert into `order` (order_id,uid,pid,price) values(#{orderId},#{uid},#{pid},(select price from product where pid=#{pid}))")
-    int createOrder(@Param("orderId") String orderId, @Param("uid") int uid, @Param("pid") int pid);
+    int createOrder(@Param("orderId") String orderId, @Param("uid") int uid, @Param("pid") long pid);
 
     //支付，修改status状态
     @Update("update `order` set status=1 where uid=#{uid} and price=#{price} and order_id=#{orderId}")
@@ -26,7 +27,7 @@ public interface OrderDao {
 
     //查询某用户某商品的订单
     @Select("select * from `order` where uid=#{uid} and pid=#{pid}")
-    List<Order> getOrderByUidAndPid(@Param("uid") int uid, @Param("pid") int pid);
+    List<Order> getOrderByUidAndPid(@Param("uid") int uid, @Param("pid") long pid);
 
     //清空订单表
     @Update("truncate table `order`")
